+++ Scripts/Update-MongoFiles.ps1
@@ -4,10 +4,14 @@
 	Updates the file system snapshot database.
 
 .Description
-	Requires: Mdbc module
 	Server: local
 	Database: test
 	Collection: files
+
+	Required modules:
+	* Mdbc: <https://github.com/nightroman/Mdbc>
+	* SplitPipeline: <https://github.com/nightroman/SplitPipeline> [1]
+	[1] SplitPipeline is needed only with the switch -Split
 
 	The script scans the specified directory tree, updates file and directory
 	documents, and then removes orphan documents that have not been updated.
@@ -30,13 +34,19 @@
 		collection. Note that for paths like C:\ it may take several minutes.
 		PowerShell Get-ChildItem is relatively slow.
 
+.Parameter Split
+		Tells to perform parallel data processing using Split-Pipeline from the
+		SplitPipeline module. Processing of massive data is typical for tasks
+		related to databases and Split-Pipeline may improve performance.
+
 .Link
 	Get-MongoFile.ps1
 #>
 
 param
 (
-	$Path = '.'
+	[Parameter()]$Path = '.',
+	[switch]$Split
 )
 
 Set-StrictMode -Version 2
@@ -60,11 +70,13 @@ if (!$Path.EndsWith('\')) {
 Import-Module Mdbc
 $collection = Connect-Mdbc . test files
 
-### Update data for existing files
+### Gets input items.
+function Get-Input {
+	Get-ChildItem -LiteralPath $Path -Force -Recurse -ErrorAction Continue
+}
 
-Write-Host "Updating data for existing files in $Path ..."
-$watch = [Diagnostics.Stopwatch]::StartNew()
-Get-ChildItem -LiteralPath $Path -Force -Recurse -ErrorAction Continue | .{process{
+### New documents from input items.
+function New-Document {process{
 	$document = New-MdbcData -DocumentId $_.FullName
 	$document.Name = $_.Name
 	$document.Extension = $_.Extension
@@ -76,20 +88,34 @@ Get-ChildItem -LiteralPath $Path -Force -Recurse -ErrorAction Continue | .{proce
 		$document.Length = $_.Length
 	}
 	$document
-}} |
-Add-MdbcData -Update $collection -ErrorAction Continue
-Write-Host $watch.Elapsed
+}}
+
+### Update data for existing files.
+$time = [Diagnostics.Stopwatch]::StartNew()
+Write-Host "Updating data for existing files in $Path ..."
+if ($Split) {
+	Import-Module SplitPipeline
+	Get-Input |
+	Split-Pipeline -Auto -Load 150 -Queue 10000 -Verbose -Module Mdbc -Function New-Document `
+	-Begin {
+		$collection = Connect-Mdbc . test files
+	} `
+	-Script {
+		$input | New-Document | Add-MdbcData -Update $collection -ErrorAction Continue
+	}
+}
+else {
+	Get-Input | New-Document | Add-MdbcData -Update $collection -ErrorAction Continue
+}
 
 ### Remove data of missing files
-
 $pattern = '^' + [regex]::Escape($Path)
 $query = query @(
 	query Updated -LT $updated
 	query _id -Match $pattern
 )
-
 Write-Host "Removing data of missing files in $Path ..."
-$watch = [Diagnostics.Stopwatch]::StartNew()
-
-Remove-MdbcData -Safe $collection $query
-Write-Host $watch.Elapsed
+$watch2 = [Diagnostics.Stopwatch]::StartNew()
+$result = Remove-MdbcData -Safe $collection $query
+Write-Host "Response: $($result.Response)"
+Write-Host $time.Elapsed